finishing off bayes chunk
